Add generic types and split into multiple files
