+++ optd-dsl/src/analyzer/mod.rs
@@ -3,5 +3,5 @@ pub mod from_ast;
 pub mod hir;
 pub mod map;
 pub mod semantic_checker;
-pub mod r#type;
+pub mod types;
 pub mod type_checker;
